Renamed basecheckbox to amplifycheckbox
