+++ packages/vue/src/index.ts
@@ -17,7 +17,7 @@ import {
   VerifyUser,
   AuthenticatorSignUpFormFields,
   BaseTextField,
-  BaseCheckBox,
+  AmplifyCheckBox,
 } from './components/index';
 
 import { useAuthenticator } from './composables/useAuth';
@@ -41,7 +41,7 @@ export default {
     app.component('VerifyUser', VerifyUser);
     app.component('ConfirmVerifyUser', ConfirmVerifyUser);
     app.component('BaseTextField', BaseTextField);
-    app.component('BaseCheckBox', BaseCheckBox);
+    app.component('AmplifyCheckBox', AmplifyCheckBox);
     app.component(
       'AuthenticatorSignUpFormFields',
       AuthenticatorSignUpFormFields
@@ -66,7 +66,7 @@ export {
   ConfirmVerifyUser,
   VerifyUser,
   BaseTextField,
-  BaseCheckBox,
+  AmplifyCheckBox,
   useAuthenticator,
 };
